+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-- Added automatic formatting code with pre-commit and clang-format
+- Changed default organ cache and organ settings directories to ones inside the GrandOrgue directory https://github.com/GrandOrgue/grandorgue/discussions/934
+- Added automatic formatting code with pre-commit and clang-format https://github.com/GrandOrgue/grandorgue/issues/918
 # 3.5.0 (2022-01-05)
 - Added default regex patterns for matching MIDI devices with alsa, jack and winmm https://github.com/GrandOrgue/grandorgue/issues/885
 - Added capability of matching MIDI devices with regex expressions https://github.com/GrandOrgue/grandorgue/issues/885